requirements.txt
awsebcli==3.0.10
botocore==1.4.25
cement==2.4.0
colorama==0.3.7
diff-match-patch==20121119
Django==1.8.3
django-excel==0.0.4
django-import-export==0.2.8
django-mathfilters==0.4.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
future==0.15.2
jmespath==0.9.0
pathspec==0.3.4
pyexcel==0.2.0
pyexcel-io==0.1.0
pyexcel-webio==0.0.6
pyexcel-xls==0.1.0
python-dateutil==2.5.3
PyYAML==3.11
requests==2.9.1
semantic-version==2.5.0
six==1.8.0
tablib==0.10.0
texttable==0.8.4
virtualenv==15.0.2
websocket-client==0.37.0
xlrd==0.9.4
XlsxWriter==0.8.4
xlwt==1.0.0
xlwt-future==0.8.0
